api/pyproject.toml
[project]
name = "api"
version = "1.0.0"
description = "A simple FastAPI service for the personal site"
authors = [
    {name = "Braelyn Boynton", email = "bboynton97@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "alembic>=1.14.1",
    "sqlalchemy>=2.0.45",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "e2b-code-interpreter>=1.0.5",
    "python-dotenv>=1.0.1",
]

[project.scripts]
start = "uvicorn main:app --host 0.0.0.0 --port 8000"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]
